Update find place to use MVC pattern
