+++ src/main/resources/fxml/find_place.fxml
@@ -1,0 +1,57 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright 2017 Esri.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License"); you may not
+  ~ use this file except in compliance with the License. You may obtain a copy of
+  ~ the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+  ~ WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+  ~ License for the specific language governing permissions and limitations under
+  ~ the License.
+  -->
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ComboBox?>
+<?import javafx.scene.layout.StackPane?>
+<?import javafx.scene.layout.VBox?>
+<?import com.esri.arcgisruntime.mapping.view.MapView?>
+<?import javafx.collections.FXCollections?>
+<?import java.lang.String?>
+<StackPane fx:controller="com.esri.samples.search.find_place.FindPlaceController"
+           xmlns:fx="http://javafx.com/fxml" stylesheets="/css/style.css">
+    <MapView fx:id="mapView"/>
+    <VBox StackPane.alignment="TOP_LEFT" spacing="5" maxWidth="250" maxHeight="50" styleClass="panel-region">
+        <StackPane.margin><Insets left="10" top="10"/></StackPane.margin>
+        <padding><Insets top="6" left="6" right="6" bottom="6"/></padding>
+        <ComboBox fx:id="placeBox" promptText="Find a place" editable="true" maxWidth="Infinity">
+            <items>
+                <FXCollections fx:factory="observableArrayList">
+                    <String fx:value="Starbucks"/>
+                    <String fx:value="Coffee"/>
+                    <String fx:value="Bank"/>
+                    <String fx:value="Gas station"/>
+                    <String fx:value="Empire State Building"/>
+                </FXCollections>
+            </items>
+        </ComboBox>
+        <ComboBox fx:id="locationBox" promptText="Near..." editable="true" maxWidth="Infinity">
+            <items>
+                <FXCollections fx:factory="observableArrayList">
+                    <String fx:value="Los Angeles, CA"/>
+                    <String fx:value="New York City, New York"/>
+                </FXCollections>
+            </items>
+        </ComboBox>
+        <Button text="Search" onAction="#search" maxWidth="Infinity"/>
+    </VBox>
+    <Button fx:id="redoButton" text="Redo search in this area" onAction="#searchByCurrentViewpoint" disable="true"
+            StackPane.alignment="BOTTOM_CENTER" maxWidth="260">
+        <StackPane.margin><Insets bottom="15"/></StackPane.margin>
+    </Button>
+</StackPane>
